Add war for deployment.

+++ input_files/import-postgres.sql
@@ -1,5 +1,5 @@
 create table Member (
-	id int,
+	id bigint,
 	name varchar(100),
 	email varchar(100),
 	phone_number varchar(100)
